utils: Move hex serializers to utils from rpc_adapter.
